refactor(project id verifying): moved from middleware to controller
Users can request deleting projects, with the project id as path
variable. Checking if they are entitled to delete the project in
question has been moved from a middleware (deleted) to the controller.

+++ TaskManager/TaskManager/Controllers/API/ProjectAPIController.cs
@@ -27,8 +27,9 @@ namespace TodoWithDatabase.Controllers.API
         public ActionResult<AssigneeWithCardsDTO> GetAssigneeProjectFor(string projectId)
         {
             var project = _projectService.GetWithAssigneeProjects(projectId);
+            var userIsEntitledToDelete = _projectService.userIsOnlyCollaboratorOnProject(User.Identity.Name, projectId);
 
-            if (project.AssigneeProjects.Count == 1)
+            if (userIsEntitledToDelete && project.AssigneeProjects.Count == 1)
             {
                 _projectService.Delete(projectId);
                 return NoContent();
